AnnualReport is now main page updated router accordingly updated unit tests accordingly
