+++ coops-ui/src/router.ts
@@ -1,6 +1,6 @@
 import Vue from 'vue'
 import Router from 'vue-router'
-import Home from './views/Home.vue'
+import AnnualReport from '@/views/AnnualReport.vue'
 import axios from '@/axios-auth'
 
 Vue.use(Router)
@@ -54,8 +54,12 @@ let router = new Router({
   routes: [
     {
       path: '/',
-      name: 'home',
-      component: Home,
+      redirect: '/annual-report'
+    },
+    {
+      path: '/annual-report',
+      name: 'annual-report',
+      component: AnnualReport,
       meta: {
         requiresAuth: true
       }
@@ -67,6 +71,11 @@ let router = new Router({
       // this generates a separate chunk (about.[hash].js) for this route
       // which is lazy-loaded when the route is visited.
       component: () => import(/* webpackChunkName: "about" */ './views/About.vue')
+    },
+    {
+      // default/fallback route
+      path: '*',
+      redirect: '/'
     }
   ]
 })
