+++ coops-ui/tests/unit/ChangeOfAddress.spec.ts
@@ -2,7 +2,7 @@ import Vue from 'vue'
 import Vuetify from 'vuetify'
 
 import App from '@/App.vue'
-import Home from '@/views/Home.vue'
+import AnnualReport from '@/views/AnnualReport.vue'
 import RegisteredOfficeAddress from '@/components/ARSteps/RegisteredOfficeAddress.vue'
 import store from '@/store'
 import sinon from 'sinon'
@@ -161,7 +161,7 @@ describe('App.vue', () => {
     let rootInstance = new RootConstructor({ store: store })
     rootvm = rootInstance.$mount()
 
-    const Constructor = Vue.extend(Home)
+    const Constructor = Vue.extend(AnnualReport)
     let instance = new Constructor({ store: store })
     vm = instance.$mount()
