+++ coops-ui/tests/unit/Directors.spec.ts
@@ -2,7 +2,7 @@ import Vue from 'vue'
 import Vuetify from 'vuetify'
 
 import App from '@/App.vue'
-import Home from '@/views/Home.vue'
+import AnnualReport from '@/views/AnnualReport.vue'
 import Directors from '@/components/ARSteps/Directors.vue'
 import store from '@/store'
 import sinon from 'sinon'
@@ -94,7 +94,7 @@ describe('Directors.vue', () => {
     let rootInstance = new RootConstructor({ store: store })
     rootvm = rootInstance.$mount()
 
-    const Constructor = Vue.extend(Home)
+    const Constructor = Vue.extend(AnnualReport)
     let instance = new Constructor({ store: store })
     vm = instance.$mount()
